corrected small deprecations (types, array dims etc) to make the tests pass with numpy 1.24

+++ wetb/fatigue_tools/rainflowcounting/peak_trough.pyx
@@ -22,7 +22,7 @@ cpdef np.ndarray[long,ndim=1] peak_trough(np.ndarray[long,ndim=1] x, int R):
     MINZO = 1
     MAXZO = 2
     ENDZO = 3
-    S = np.zeros(x.shape[0] + 1, dtype=np.int)
+    S = np.zeros(x.shape[0] + 1, dtype=np.int32)
 
     L = x.shape[0]
     goto = BEGIN
